Added mongodb server 🌿
Planning on switching out the current sqlite based database system to
mongodb, for this reason I have added a mongodb docker file to
docker-comose and set up a database network where only the teahaz-api
container is allowed to communicate to the database.

I have bought in all these security measures because I dont expect the
average server owner to have to bother with changing database passwords,
and setting them to something secure. Instead I wont have the database
accessible by default, and ill set defaults for the database password.

note: the password should still be changable in the future, and stored
in evn for extra security.