First bit of work on this architecture re-design.
I have opted to go for a half re-write method for this refactor. I have
created a new branch with none of the original code in it. I will be
mostly pasting in and modifying to old code as its not all bad and I
don't want to spend years on this re-factor.